초기화면 수정
