+++ AppleKiwi/AppleKiwi/Assets.xcassets/kiwi image/sprout kiwi.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "새싹 키위.png",
+      "filename" : "새싹 키위.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
